Daniela

+++ frontend/src/app/pages/edit-event/edit-event.page.ts
@@ -6,6 +6,8 @@ import { ModalController } from '@ionic/angular';
 import { LocationSelectPage } from '../location-select/location-select.page';
 import { AlertController } from '@ionic/angular';
 import { TranslateService } from '@ngx-translate/core';
+import { DeleteAlertService } from 'src/app/services/delete-alert.service';
+import { StorageService } from 'src/app/services/storage.service';
 
 @Component({
   selector: 'app-edit-event',
@@ -21,7 +23,7 @@ export class EditEventPage implements OnInit {
   ListConv: any[];
   numConv = 0;
   convocats = [];
-  membershipTeam;
+  membershipTeam = [];
 
   locationForm = this.formBuilder.group({
     latitude: [null, [Validators.required]],
@@ -48,7 +50,9 @@ export class EditEventPage implements OnInit {
     private route: ActivatedRoute,
     private router: Router,
     private alertCtrl: AlertController,
-    private translateService: TranslateService
+    private translateService: TranslateService,
+    private deleteAlert: DeleteAlertService,
+    private storageServ: StorageService
   ) { 
     this.route.queryParams.subscribe(params => {
       this.evId = this.router.getCurrentNavigation().extras.state.evId;
@@ -92,25 +96,34 @@ export class EditEventPage implements OnInit {
     return modal.present();
   }
 
-  deleteEvent() {
-    console.log(this.evId);
-    this.apiProv.deleteEvent(this.evId)
-    .then(() => {
-      this.router.navigate(['/calendar']);
+  async deleteEvent() {
+    this.deleteAlert.showConfirm(this.event.type, this.event.title).then((res) => {
+      if(res) {
+        if(this.event.type == 'match') {
+          let teamId = this.apiProv.getTeamId();
+          this.storageServ.deleteEventFiles(teamId, this.evId);
+        }
+        this.apiProv.deleteEvent(this.evId)
+        .then(() => {
+          this.router.navigate(['/calendar']);
+        });
+        
+      }
     });
+    
   }
 
   onDone() {
     if(this.event.type == "match") {
       this.apiProv.editMatch(this.editEventForm.value)
       .then(() => {
-        this.router.navigate(['/event', this.editEventForm.get('eventId').value]);
+        this.router.navigate(['/event', this.evId]);
       });
     }
     else {
       this.apiProv.editTraining(this.editEventForm.value)
       .then(() => {
-        this.router.navigate(['/event', this.editEventForm.get('eventId').value]);
+        this.router.navigate(['/event', this.evId]);
       });
     }
   }
@@ -147,7 +160,7 @@ export class EditEventPage implements OnInit {
               handler : () => {
                 this.apiProv.createCall(this.evId, this.convocats)
                 .then(() => {
-                  this.router.navigate(['/event', this.editEventForm.get('eventId').value]);
+                  this.router.navigate(['/event', this.evId]);
                 });
               }
             }
@@ -162,7 +175,7 @@ export class EditEventPage implements OnInit {
     for (let member of this.membershipTeam) {
       if (member.isChecked) {
         ++this.numConv;
-        this.convocats.push(member.userId);
+        this.convocats.push(member.id);
       }
     }
     this.presentAlert();
@@ -171,20 +184,26 @@ export class EditEventPage implements OnInit {
   getMembersTeam() {
     this.apiProv.getMembers()
       .then((data) => {
-        this.membershipTeam = data;
-        if (this.membershipTeam.length == 0) this.membershipTeam = null;
-        else {
-          for ( let mem of this.membershipTeam ) {
-            mem.isChecked = false;
+      let tmp: any = data;
+        tmp.forEach(element => {
+          if(element.type == 'player') {
+            this.apiProv.getUser(element.userId).subscribe((info: any) => {
+              let player = {
+                id: element.userId,
+                name: info.userName,
+                isChecked: false
+              }
+              this.membershipTeam.push(player);
+            })
           }
-        }
-      }); 
+        });
+    }); 
   }
 
   listPlayers() {
     this.getMembersTeam();
-    if ( this.membershipTeam != null ) {
-      if ( this.ListConv != null ) {
+    if ( this.membershipTeam.length != 0 ) {
+      if ( this.ListConv.length != 0 ) {
         this.anySelected = true;
         for ( let mem of this.membershipTeam ) {
           for ( let conv of this.ListConv ) {
